requirements.txt
requests>=2.31.0
psycopg2-binary>=2.9.10
python-dotenv>=1.0.0
pandas>=2.2.3
sqlalchemy>=2.0.25
tenacity>=8.2.3
numpy>=2.1.0
scikit-learn>=1.5.0
